Dummy registerInterface; Move registry to registry.py
