+++ src/plone.server/plone/server/registry.py
@@ -1,7 +1,65 @@
 # -*- coding: utf-8 -*-
+from BTrees._OOBTree import OOBTree
+from persistent.mapping import PersistentMapping
 from plone.server import _
+from plone.server.interfaces import IRegistry
 from zope import schema
+from zope.interface import alsoProvides
+from zope.interface import implementer
 from zope.interface import Interface
+
+
+class RecordsProxy(object):
+    """A adapter that knows how to store data in registry.
+
+    Each value will be stored as a primitive in the registry under a key
+    that consists of the full dotted name to the field being stored.
+
+    This class is not sufficient as an adapter factory on its own. It must
+    be initialised with the schema interface in the first place. That is the
+    role of the Annotations factory below.
+    """
+
+    def __init__(self, context, iface, prefix=None):
+        self.__dict__['records'] = context
+        self.__dict__['schema'] = iface
+        if prefix is not None:
+            self.__dict__['prefix'] = prefix + '.'
+        else:
+            self.__dict__['prefix'] = iface.__identifier__ + '.'
+        alsoProvides(self, iface)
+
+    def __getattr__(self, name):
+        if name not in self.__dict__['schema']:
+            raise AttributeError(name)
+
+        records = self.__dict__['records']
+        key_name = self.__dict__['prefix'] + name
+        if key_name not in records:
+            return self.__dict__['schema'][name].missing_value
+
+        return records[key_name]
+
+    def __setattr__(self, name, value):
+        if name not in self.__dict__['schema']:
+            super(RecordsProxy, self).__setattr__(name, value)
+        else:
+            prefixed_name = self.__dict__['prefix'] + name
+            self.__dict__['records'][prefixed_name] = value
+
+
+@implementer(IRegistry)
+class Registry(PersistentMapping):
+
+    def __init__(self):
+        self._data = OOBTree()
+        super(Registry, self).__init__()
+
+    def forInterface(self, iface, check=True, omit=(), prefix=None):
+        return RecordsProxy(self, iface, prefix=prefix)
+
+    def registerInterface(self, iface, omit=(), prefix=None):
+        pass
 
 
 ACTIVE_AUTH_EXTRACTION_KEY = \
